Compute normalized probabilities.

Compute normalized probabilities from unnormalized log-probabilities.

normalizelogweights(logw)

Arguments

logw
Vector of unnormalized log-probabilities.

Details

Guards against underflow or overflow by adjusting the log-probabilities so that the largest probability is 1.

Value

Normalized probabilities such that the sum is equal to 1.

References

P. Carbonetto and M. Stephens (2012). Scalable variational inference for Bayesian variable selection in regression, and its accuracy in genetic association studies. Bayesian Analysis 7, 73--108.

Examples

  logw <- rnorm(6)
  w    <- normalizelogweights(logw)