Summarize a fitted variable selection model.

Generate a summary of the Bayesian variable selection model fitted using variational approximation methods.

# S3 method for varbvs
summary(object, cred.int = 0.95, nv = 5, nr = 1000, …)
  # S3 method for summary.varbvs
print(x, digits = 3, …)

Arguments

object
Output of function varbvs.

cred.int
Size of credible interval (number between 0 and 1).

nv
Show detailed statistics for top nv variables, ranked according to their posterior inclusion probabilities.

nr
Number of Monte Carlo samples to draw to estimate credible intervals for coefficients of selected variables.

x
Output of function summary.varbvs.

digits
Number of digits shown when printing posterior probabilities of top nv variables.

...
Additional print arguments.

Details

The printed summary is divided into three parts. The first part summarizes the data and optimization settings. It also reports the hyperparameter setting that yields the largest marginal likelihood---more precisely, the approximate marginal likelihood computed using the variational method. For the linear regression only (family = "gaussian"), it reports the estimated proportion of variance in the outcome explained by the model (PVE), and the credible interval of the PVE estimate brackets. Note that this is the PVE in the outcome after removing variance in the outcome due to linear effects of the covariates.

The second part summarizes the approximate posterior distribution of the hyperparameters (sigma, sa, logodds). The "estimate" column is the value averaged over hyperparameter settings, treating objectlogw as (unnormalized) log-marginal probabilities. The next column, labeled "Pr>x", where x = cred.int gives the credible interval based on these weights (computed using function cred).

The third part summarizes the variable selection results. This includes the total number of variables included in the model at different posterior probability thresholds, and a more detailed summary of the variables included in the model with highest posterior probability. For family = "gaussian", the "PVE" column gives the estimated proportion of variance in the outcome explained by the variable (conditioned on being included in the model). Again, this is the PVE after removing variance in the outcome due to linear effects of the covariates. Finally, note that the credible intervals reported in the right-most column are Monte Carlo estimates, so the interval will be slightly different each time summary is called; a more accurate estimate can be obtained by setting input nr to a larger number, at the cost of increased computation time.

Value

An object of class summary.varbvs, to be printed by print.summary.varbvs.

References

P. Carbonetto and M. Stephens (2012). Scalable variational inference for Bayesian variable selection in regression, and its accuracy in genetic association studies. Bayesian Analysis 7, 73--108.

See also

varbvs, varbvscoefcred

Examples

  # See help(varbvs) for examples.